Maintenance

+++ Examples/Contract_Crossover.sol
@@ -1,12 +1,12 @@
 pragma solidity>0.8.0;//SPDX-License-Identifier:None
 
 interface IOwlContract{
-    function MINT(address _t,uint256 _a)external;
+    function MINT(address _t,uint _a)external;
 } 
 
 contract testCrossContract{
     IOwlContract private ioc;
-    function testMint(uint256 _t)external{
+    function testMint(uint _t)external{
         ioc.MINT(msg.sender,_t);
     }
     function setAddress(address _a)external{
